Updated bed area and start gcode
These changes made to match the actual size of the printbed which is 235x235, printable area is still the same. Start G-code primes nozzle at negative x value to compensate. If you are still using an old build-plate model, it will show as printing off the build-plate as the actual size of the v3se is 235x235 with a 220x220 printable area (meaning there is a 7.5mm border around which we can actually print on if you give negative GCODE coords)